Refactor: change material interval type to double for later resource spawn changes

+++ api/src/main/java/org/screamingsandals/bedwars/api/game/ItemSpawnerType.java
@@ -40,7 +40,7 @@ public interface ItemSpawnerType {
     /**
      * @return
      */
-    int getInterval();
+    double getInterval();
 
     /**
      * @return
